[ARRISEOS-40758] Use Widget bounds for PlatformScreen rectangle information
This is a backport of WebKit change described below. It is neeeded for
BritBox, which  uses |window.screen.availHeight| JS API to make sure that
VirginMedia app version will fit into the screen. If returned value is less
than 1080, it falls back to the FVP app version with native DASH playback.

https://bugs.webkit.org/show_bug.cgi?id=193190

Provide a better screen area estimate in screenRect() and
screenAvailableRect() return values than the current 1240x640 value by
using the Widget's bounds rectangle.

This approach is only factually correct when the Widget-associated view
is displayed in fullscreen, but it provides a better estimate even when
displayed in any other case as well. WPE doesn't provide specific API
that could enable the embedding environment to provide this information.

* platform/wpe/PlatformScreenWPE.cpp:
(WebCore::screenRect): Return bounds of the Widget object.
(WebCore::screenAvailableRect): Relay the call to screenRect().