[ARRISEOS-40859][SoupNetworkSession] Drop open connections on network change
This is a downstream of:
https://github.com/WebPlatformForEmbedded/WPEWebKit/commit/43877edc4454aa4c7e4073aa6658cedfb62756f1

Libsoup internally creates sockets and tries to keep them alive for the sake
of incoming request to the same host. It works just fine, unless the network
change occurs in the meantime. In such case, libsoup does not drop the active
connection automatically.

This change introduces network change detection, which triggers
soup_session_abort and closes the hanged sockets.