[ARRISAPOL-2585] Treat whitelisted websockets as secure
Onesdk userscript library connects to unencrypted websockets, causing
browser to mark whole document as mixed content. For such documents,
browser doesn't allow access to cookies that are marked as "secure".

BBC sounds application uses secure cookie to store selected user
profile. Without it, it's not possible to switch to different profile.

In ARRISEOS-29974 whitelist was added so such connection can be made in
security strict mode. This commit extends that behaviour and those
connections are treated as secure.

All whitelisted connections are with localhost, so no unencrypted data
leaves device.

Commit not eligible for upstream.