[ARRISEOS-41099] Repaint page after receiving wpe_compositor_refresh_event
The entire page should be repainted after receiving
wpe_compositor_refresh_event. This will make webbrowser repond
to visibility change from awc_cli.

Event flow:
We receive an information about handle_configure method
execution in wpe-backend-rdk. Thanks to this execution we know
that the compositor change state and we need to repaint the
app.
As we get this info in network process thread we send it to the UI
thread in wpe-backend-rdk via
Wayland::EventDispatcher::singleton().sendEvent(). The wpe-browser
is attached to EventDispatcher so by IPC we get this info in
WPEView.cpp in wpe-webkit, wpe_view_backend_input_client callback.
This event is received in UI thread while WebPage is working on
network process thread. We send the event to the webpage by:
page.repaintAfterCompositorReconfigure(); When we receive event on
the proper thread, process and proper class we force the app to be
repainted. To this we execute: m_drawingArea->forceRepaint(true);