[ARRISEOS-43575]: AppleTV - enable switch from promo content to main content (#225)
Promo video at the beginning of playback is not encrypted.
Encrypted playback after that promo is presnt on the same MSE blob.
On AppendPipeline level we remove from qtdemux level promo pad (eg. video_0)
and create main video pad (eg. video_1), which needs to be connected to appsink.

This is a backport of upstream fix:
https://github.com/WebPlatformForEmbedded/WPEWebKit/pull/862

Co-authored-by: Jacek Skiba <jacek.skiba@consult.red>