[ARRISEOS-43508]: Drop float precision in surface comparison
This commit fixes a crash that was encountered in Pixman. The crash was
caused by a subsurface that is bigger than the surface, so we end up
reading not allocated memory.
In the committed code, we change FloatRect type to IntRect. Thanks to
this we prevent of IntRect promotion to FloatRect from cairoSurfaceSize,
so with the fix we compare int size not floats. Float comparisons are
untrustworthy.
In a crash scenario the code state that 480.000031 is not equal to 480
and we have to create a subsurface.
enclosingIntRect method used ceil method to set width, and height,
because of this 480.000031 becomes 481 which is bigger than the
original surface.