[HUMAXEOS-4885]: Fix artifacts on STV at the end of the movie
Configure sink to not drop buffers which are outside the current segment.
This way we do not lost video frames which are valid.

Example of currently dropped buffer:
- gst segment stop time: 1:30:00.00
- gst buffer dts: 1:29:59.80, pts: 1:30:00.04
Current buffer is dropped because pts (1:30:00.04) is higher then gst stop time (1:30:00.00).

This condition is too strict for STV content. We don't see also that strict requirement in Dash specification.