[ONEM-30000]: Disable clipping mechanism in gst for all apps
Configure sink to not drop buffers which are outside the current segment.
This way we do not lost video frames which are valid.

Example of currently dropped buffer:
- gst segment stop time: 1:30:00.00
- gst buffer dts: 1:29:59.80, pts: 1:30:00.04
Current buffer is dropped because pts (1:30:00.04) is higher then gst stop time (1:30:00.00).

Dash specification allows such streams.

In HUMAXEOS-4885 we already disabled clipping mechanism, but this task
did this only in STV application.