Fixes for apple tv (#288)
* [GStreamer][MSE] Append pipeline is counting streams incorrectly with new init segments
https://bugs.webkit.org/show_bug.cgi?id=254664

Reviewed by Alicia Boya Garcia.

When you append another init segment we match GStreamer pads against tracks but we are not counting properly as we are
skipping the pads that are already linked we can end up with less pads than tracks. Besides it is against the spec. We
need to match the number of tracks and track IDs if there are more than one for a kind.

* LayoutTests/media/media-source/media-source-seek-detach-crash.html: Fail the test earlier instead of waiting for a
timeout.
* LayoutTests/platform/glib/TestExpectations: Unmark
media/encrypted-media/encrypted-media-append-encrypted-unencrypted.html as failure. It works now.
* Source/WebCore/platform/graphics/gstreamer/mse/AppendPipeline.cpp:
(WebCore::AppendPipeline::didReceiveInitializationSegment):
(WebCore::AppendPipeline::recycleTrackForPad):

Canonical link: https://commits.webkit.org/264676@main

* [ONEM-31328] Ignore unknown tracks in AppendPipeline

* [ONEM-31328] Dynamically attach decryptor

Probe has been added to src pad of WebKitMediaSrc. Based on incomming
caps, decryptor is either attached or detached. This fixes playback of
content Apple TV, when begining of the video is unencrypted, then it's
switched to encrypted.

Based on b766c8a8dcca51495d18b775554be7318bb1bf30

---------

Co-authored-by: Xabier Rodriguez-Calvar <calvaris@igalia.com>