ONEM-33416 Add resolution limit for h.264 codec on EOS platform
This change is based on a similar pull request done for WPE 2.22
https://github.com/LibertyGlobal/WPEWebKit/pull/258/files
For WPE 2.38 only a part of that pull request was taken that
specifically addresses EOS platform.

In WPE 2.38 the westeros-sink element is used in gstreamer pipeline
that also requires limits on stream resolution. The complementary
fix that addresses westeros-sink can be found here:
https://gerrit.onemw.net/c/meta-lgi-wpe/+/114656

The solution introduced in this commit could not be implemented
in the same way as the one for westeros-sink because it requires
to be only applied in case when h.264/avc codec was used and the
mechanism based on the config variable VIDEO_DECODING_LIMIT
supports only resolution and framerate limits without checking
what codec is in use.

Preferably the VIDEO_DECODING_LIMIT implementation should be
extended to support filtering narrowed to a specified codec.
This change could then be eliminated as required limits
configuration could be done in a single place,
in VIDEO_DECODING_LIMIT variable.