ARRISAPOL-3614 wpe: fix in GstState handling
MediaPlayerPrivateGStreamer::changePipelineState assumes that
when pipeline state change is requested, and the requested state
is same as current state, that request can be rejected. What was
not taken into account is that even if newState == currentState,
there might still be some other state pending, and the pipeline
state might still need to be updated in such case.