ONEM-39602 resize FBOs on suspend
A PoC example of resizing webgl canvas backing
FBOs, to decrease GFX memory usage when suspending.

Seems that fully dropping ('reshape(0,0)') the fbos didn't
work, but with reshape(16,16), it is possible to redraw the
screen at the resume. After resume it is still necessary
to force screen redraw (eg with some navigation); this
probably would need to be handled somehow.

With 1920x1080 res, this change decreases GFX mem usage by
around 32MB when is suspend. (tested on Wikipedia app), provided
the application is using canvas/wegbl at all (which seems to be
 the case for most of the lightning apps)